Adapt LibTest/async for strong mode.

+++ LibTest/async/Stream/distinct_A01_t01.dart
@@ -44,5 +44,5 @@ main() {
   check([1, 2, 2, 3]);
   check([1, 2, null, null]);
   check(new Iterable.generate(0, (int index) => 1));
-  check(new Iterable.generate(10, (int index) => [0]));
+  check(new Iterable.generate(10, (int index) => [0] as int));
 }
